Reestructuracion Miercoles 23/11/2016
Combine ambos proyectos.

+++ database/migrations/2016_10_13_000617_add_create_patrimonios_table.php
@@ -20,7 +20,7 @@ class AddCreatePatrimoniosTable extends Migration
             $table->foreign('idEmpresa')->references('id')->on('organizaciones')->onDelete('cascade');
 
             $table->integer('anio');
-            $table->integer('valor');
+            $table->bigInteger('valor');
             $table->timestamps();
         });
     }
